+++ CHANGELOG.rst
@@ -9,6 +9,19 @@ and this project adheres to `Semantic Versioning`_.
 .. _Keep a Changelog: https://keepachangelog.com/en/1.0.0/
 .. _Semantic Versioning: https://semver.org/spec/v2.0.0.html
 
+[0.0.3] - 2025-08-24
+--------------------
+
+Added
+^^^^^
+- Determine minimum Disparity my approximate distance function added.
+- Pick point from point cloud to display (x,y,x) values in terminal added.
+- Origin axis frame added to pointcloud visualisation.
+- Live calibration function added.
+- Example file added for live calibration.
+- Added stereo camera detection by name function.
+- Added rectification of coordinates from camera parameters.
+
 [0.0.2] - 2025-03-06
 --------------------